from stefan
git-svn-id: https://qucs.svn.sourceforge.net/svnroot/qucs/trunk@1014 b5b04e8c-4942-46c9-ab4f-83783d557d1c

+++ qucs/qucs/components/component.cpp
@@ -735,7 +735,7 @@ bool Component::load(const QString& _s)
     tmp = 2;
   else tmp = counts;
 
-  for(; tmp<=counts/2; tmp++)
+  for(; tmp<=(int)counts/2; tmp++)
     Props.append(new Property("p", "", true, " "));
 
   // load all properties
